+++ test/intrinsics.jl
@@ -106,3 +106,22 @@ end
 @test unsafe_load(Ptr{Nothing}(0)) === nothing
 struct GhostStruct end
 @test unsafe_load(Ptr{GhostStruct}(rand(Int))) === GhostStruct()
+
+
+@test Core.Intrinsics.atomics_fence(:seq_cst) === nothing
+@test Core.Intrinsics.atomics_pointerref(C_NULL, :seq_cst) == nothing
+let r = Ref{Int}(10)
+    p = Base.unsafe_convert(Ptr{Int}, r)
+    GC.@preserve r begin
+        @test Core.Intrinsics.atomics_pointerref(p, :seq_cst) === 10
+        @test Core.Intrinsics.atomics_pointerset(p, 1, :seq_cst) === p
+        @test Core.Intrinsics.atomics_pointerref(p, :seq_cst) === 1
+        @test Core.Intrinsics.atomics_pointercmpxchg(p, 100, 1, :seq_cst) === true
+        @test Core.Intrinsics.atomics_pointerref(p, :seq_cst) === 100
+        @test Core.Intrinsics.atomics_pointercmpxchg(p, 1, 1, :seq_cst) === false
+        @test Core.Intrinsics.atomics_pointerref(p, :seq_cst) === 100
+        @test Core.Intrinsics.atomics_pointerop(p, 1, +, :seq_cst) == 100
+        @test Core.Intrinsics.atomics_pointerop(p, 1, +, :seq_cst) == 101
+        @test Core.Intrinsics.atomics_pointerref(p, :seq_cst) == 102
+    end
+end